viaseguraapi/scripts/requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
python-dateutil==2.8.2